fix: remove gas workaround
